+++ lib/llm/Cargo.toml
@@ -44,6 +44,7 @@ harness = false
 [dependencies]
 # repo
 dynamo-runtime = { workspace = true }
+dynamo-tokens = { workspace = true }
 
 # workspace
 anyhow = { workspace = true }
